Better handle strs when bytes are expected internally.
The Python half of the library passes str (PyUnicode) objects to the C half for
authzid and mech when dealing with SASLObjects. The C half expected these to be
bytes (PyBytes) objects and attempted to convert them to C strings without
checking. This now checks whether each of these objects are PyUnicodes and
encodes them as UTF-8 to create PyBytes objects before continuing as before.
It's UTF-8 as per the SASL spec.